adapt works

+++ menu/src/main/resources/META-INF/migrate/20140819094130_create_menu_items.sql
@@ -2,8 +2,8 @@
 -- Migration SQL that makes the change goes here.
 
 CREATE TABLE IF NOT EXISTS menu_items (
-  id           INT(10) UNSIGNED NOT NULL PRIMARY KEY AUTO_INCREMENT,
-  parent_id    INT(10) UNSIGNED,
+  id          BIGINT(10) UNSIGNED NOT NULL PRIMARY KEY AUTO_INCREMENT,
+  parent_id   BIGINT(10) UNSIGNED,
   name         VARCHAR(100)     NOT NULL,
   state        VARCHAR(100)     NOT NULL UNIQUE,
   css          VARCHAR(100),
